requirements.txt
airflow==1.7.1.3
psycopg2==2.6.2
pytest==3.0.5
testing.postgresql==1.3.0
